fixup another buffersize typecast

+++ src/tool_operate.c
@@ -926,7 +926,7 @@ static CURLcode config2setopts(struct GlobalConfig *global,
       curl_off_t num;
       const char *p = env;
       if(!curlx_str_number(&p, &num, LONG_MAX))
-        my_setopt_long(curl, CURLOPT_BUFFERSIZE, num);
+        my_setopt_long(curl, CURLOPT_BUFFERSIZE, (long)num);
     }
     else
 #endif
